class controle de corridas entity refactored

+++ CityGusa-api/src/main/java/com/citygusa/com/citygusaapi/Dto/ControleOperacionalDto.java
@@ -34,7 +34,7 @@ public class ControleOperacionalDto {
     private Integer umidade;
     private BigDecimal umidadeMedia;
     private Integer gusaKg;
-    private Double acumuladoKilos;
+    private BigDecimal acumuladoKilos;
     private Integer acumuladoCarga;
     private Integer acumuladoCargaSeca;
     private BigDecimal mediaHoraCarga;
